backend/pyproject.toml
[project]
name = "seedr-backend"
version = "0.1.0"
description = "Seedr hydroponic scheduler"
readme = "README.md"
authors = [{name = "Seedr"}]
requires-python = ">=3.11"
dependencies = [
    "django-ninja>=1.1",
    "uvicorn[standard]>=0.27",
    "apscheduler>=3.10",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
]

[tool.setuptools.packages.find]
include = ["app*", "seedr*"]
exclude = ["tests*"]

[tool.ruff]
line-length = 100
src = ["apps", "tests"]
